autoformat: move AutoformatChange model into landing_job module
It only relates to LandingJob and is too small to warrant its own module.
